ふっとび修正

+++ TGS/ProjectData/GameSources/MyUI.h
@@ -258,7 +258,7 @@ namespace basecross {
 
 		void Draw();
 
-		int GetMaxIndex() {
+		float GetMaxIndex() {
 			return m_vertex.x / m_cutOut.x;
 		}
 	};
